refactor(ui): remove IconButton; group Toolbar/FloatingPanel under components/graph
IconButton only re-defaulted Button (ghost/sm/square + required aria-label);
replaced all usages with Button directly. Moved the canvas-oriented Toolbar and
FloatingPanel primitives into components/graph/ to separate them from the regular
primitives.
