+++ frontend/packages/ui/package.json
@@ -9,9 +9,8 @@
     "./card": "./src/components/card/card.tsx",
     "./modal": "./src/components/modal/modal.tsx",
     "./styles.css": "./src/index.css",
-    "./icon-button": "./src/components/icon-button/icon-button.tsx",
-    "./toolbar": "./src/components/toolbar/toolbar.tsx",
-    "./floating-panel": "./src/components/floating-panel/floating-panel.tsx",
+    "./toolbar": "./src/components/graph/toolbar/toolbar.tsx",
+    "./floating-panel": "./src/components/graph/floating-panel/floating-panel.tsx",
     "./use-dismiss": "./src/hooks/use-dismiss.ts"
   },
   "scripts": {
