+++ frontend/packages/ui/src/components/graph/floating-panel/floating-panel.stories.tsx
