+++ frontend/packages/ui/src/components/graph/floating-panel/floating-panel.test.tsx
